requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
plaid-python==29.1.0
python-dotenv==1.1.0
pandas==2.2.3
numpy==2.2.4
scikit-learn==1.6.1
matplotlib==3.10.1
gunicorn
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9